move the details to a separate component
